# crates/api/openapi.json
{
  "components": {
    "schemas": {
...
      "GithubAccountSummary": {
        "description": "Account header on the dashboard response (port of `GithubAccountSummaryT`).",
        "properties": {
          "connected": {
            "type": "boolean"
          },
          "connectedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "login": {
            "type": [
              "string",
              "null"
            ]
          },
          "scope": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "connected"
        ],
        "type": "object"
      },
      "GithubApprovalState": {
        "enum": [
          "approved",
          "changes_requested",
          "review_required",
          "unknown"
        ],
        "type": "string"
      },
...
      "GithubCheckState": {
        "enum": [
          "success",
          "failure",
          "pending",
          "skipped",
          "neutral",
          "cancelled",
          "timed_out",
          "action_required",
          "unknown"
        ],
        "type": "string"
      },
      "GithubConnectionSummary": {
        "properties": {
          "connected": {
            "type": "boolean"
          },
          "connectedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastFour": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastUsedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastValidatedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "login": {
            "type": [
              "string",
              "null"
            ]
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "selectedRepos": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "tokenKind": {
            "type": [
              "string",
              "null"
            ]
          },
          "validationError": {
            "type": [
              "string",
              "null"
            ]
          },
          "validationStatus": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "connected",
          "selectedRepos"
        ],
        "type": "object"
      },
...
      "GithubDashboard": {
        "description": "The full dashboard response (port of `GithubDashboardT`).",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/GithubAccountSummary"
          },
          "queuePulls": {
            "items": {
              "$ref": "#/components/schemas/GithubPullRequestQueue"
            },
            "type": "array"
          },
          "queues": {
            "items": {
              "$ref": "#/components/schemas/GithubQueueCount"
            },
            "type": "array"
          }
        },
        "required": [
          "account",
          "queues",
          "queuePulls"
        ],
        "type": "object"
      },
      "GithubDashboardActor": {
        "properties": {
          "avatarUrl": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "login",
          "avatarUrl",
          "url"
        ],
        "type": "object"
      },
      "GithubDashboardLabel": {
        "properties": {
          "color": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "color"
        ],
        "type": "object"
      },
      "GithubDashboardRepository": {
        "properties": {
          "actionsUrl": {
            "type": "string"
          },
          "defaultBranch": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isPrivate": {
            "type": "boolean"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "fullName",
          "url",
          "actionsUrl",
          "isPrivate",
          "isArchived",
          "defaultBranch"
        ],
        "type": "object"
      },
...
      "GithubPullEnrichmentResult": {
        "description": "A batch enrichment entry: the ref echoed back with its enrichment (port of\n`GithubPullEnrichmentResultT`).",
        "properties": {
          "enrichment": {
            "$ref": "#/components/schemas/GithubPullRequestEnrichment"
          },
          "number": {
            "format": "int64",
            "type": "integer"
          },
          "owner": {
            "type": "string"
          },
          "repo": {
            "type": "string"
          }
        },
        "required": [
          "owner",
          "repo",
          "number",
          "enrichment"
        ],
        "type": "object"
      },
      "GithubPullRef": {
        "description": "One PR to enrich in a batch request (port of `GithubPullRefT`).",
        "properties": {
          "number": {
            "format": "int64",
            "type": "integer"
          },
          "owner": {
            "type": "string"
          },
          "repo": {
            "type": "string"
          }
        },
        "required": [
          "owner",
          "repo",
          "number"
        ],
        "type": "object"
      },
      "GithubPullRequestBasic": {
        "properties": {
          "assignees": {
            "items": {
              "$ref": "#/components/schemas/GithubDashboardActor"
            },
            "type": "array"
          },
          "author": {
            "$ref": "#/components/schemas/GithubDashboardActor"
          },
          "comments": {
            "format": "int64",
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/GithubDashboardLabel"
            },
            "type": "array"
          },
          "number": {
            "format": "int64",
            "type": "integer"
          },
          "repository": {
            "$ref": "#/components/schemas/GithubDashboardRepository"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "repository",
          "number",
          "title",
          "url",
          "author",
          "assignees",
          "labels",
          "comments",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "GithubPullRequestEnrichment": {
        "description": "Detail-derived fields, fetched lazily per PR (port of\n`GithubPullRequestEnrichmentT`).",
        "properties": {
          "additions": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "approvalState": {
            "$ref": "#/components/schemas/GithubApprovalState"
          },
          "baseRef": {
            "type": [
              "string",
              "null"
            ]
          },
          "blockerSummary": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "branchBehind": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "changedFiles": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "deletions": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "draft": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "headRef": {
            "type": [
              "string",
              "null"
            ]
          },
          "headSha": {
            "type": [
              "string",
              "null"
            ]
          },
          "latestRun": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GithubWorkflowRunSummary"
              }
            ]
          },
          "mergeable": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "mergeableState": {
            "type": [
              "string",
              "null"
            ]
          },
          "readinessBadges": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "repository": {
            "$ref": "#/components/schemas/GithubDashboardRepository"
          },
          "requestedReviewers": {
            "items": {
              "$ref": "#/components/schemas/GithubRequestedReviewer"
            },
            "type": "array"
          },
          "requiredChecks": {
            "items": {
              "$ref": "#/components/schemas/GithubRequiredCheck"
            },
            "type": "array"
          },
          "reviewComments": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "repository",
          "approvalState",
          "requestedReviewers",
          "requiredChecks",
          "readinessBadges",
          "blockerSummary"
        ],
        "type": "object"
      },
      "GithubPullRequestQueue": {
        "properties": {
          "incompleteResults": {
            "type": "boolean"
          },
          "key": {
            "$ref": "#/components/schemas/GithubQueueKey"
          },
          "label": {
            "type": "string"
          },
          "pullRequests": {
            "items": {
              "$ref": "#/components/schemas/GithubPullRequestBasic"
            },
            "type": "array"
          },
          "totalCount": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "label",
          "totalCount",
          "incompleteResults",
          "pullRequests"
        ],
        "type": "object"
      },
      "GithubQueueCount": {
        "properties": {
          "incompleteResults": {
            "type": "boolean"
          },
          "key": {
            "$ref": "#/components/schemas/GithubQueueKey"
          },
          "label": {
            "type": "string"
          },
          "totalCount": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "label",
          "totalCount",
          "incompleteResults"
        ],
        "type": "object"
      },
      "GithubQueueKey": {
        "enum": [
          "assigned",
          "review_requested",
          "authored",
          "mentioned",
          "failing_ci"
        ],
        "type": "string"
      },
...
      "GithubRepoOption": {
        "description": "One selectable repository (port of `GithubRepoOptionT`).",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isPrivate": {
            "type": "boolean"
          }
        },
        "required": [
          "fullName",
          "isPrivate",
          "isArchived"
        ],
        "type": "object"
      },
...
      "GithubRequestedReviewer": {
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/GithubRequestedReviewerKind"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "url",
          "kind"
        ],
        "type": "object"
      },
      "GithubRequestedReviewerKind": {
        "enum": [
          "user",
          "team"
        ],
        "type": "string"
      },
      "GithubRequiredCheck": {
        "properties": {
          "name": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/GithubCheckState"
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "name",
          "state",
          "required"
        ],
        "type": "object"
      },
...
      "GithubWorkflowRunSummary": {
        "properties": {
          "conclusion": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "event": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "runNumber": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "url",
          "runNumber",
          "event",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "PullsBody": {
        "properties": {
          "refs": {
            "items": {
              "$ref": "#/components/schemas/GithubPullRef"
            },
            "type": "array"
          }
        },
        "required": [
          "refs"
        ],
        "type": "object"
      },
...
      "RepoSelection": {
        "description": "`GET /me/github/repos` response.",
        "properties": {
          "available": {
            "items": {
              "$ref": "#/components/schemas/GithubRepoOption"
            },
            "type": "array"
          },
          "nextPage": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "selected": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "available",
          "selected"
        ],
        "type": "object"
      },
      "ReposBody": {
        "properties": {
          "repos": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "repos"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Workflow",
      "url": "https://github.com/"
    },
...
    "license": {
      "name": "Proprietary"
    },
    "title": "Workflow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me/github/dashboard": {
      "get": {
        "operationId": "githubDashboard",
        "parameters": [
          {
            "description": "Queue tab (default assigned)",
            "in": "query",
            "name": "tab",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubDashboard"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/github/dashboard?tab= — projected dashboard (counts + active queue).",
        "tags": [
          "github"
        ]
      }
    },
...
    "/api/me/github/pull": {
      "get": {
        "operationId": "githubPull",
        "parameters": [
          {
            "description": "Repo owner",
            "in": "query",
            "name": "owner",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Repo name",
            "in": "query",
            "name": "repo",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "PR number",
            "in": "query",
            "name": "number",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubPullRequestEnrichment"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/github/pull?owner&repo&number — enrich a single PR.",
        "tags": [
          "github"
        ]
      }
    },
...
    "/api/me/github/pulls/enrich": {
      "post": {
        "operationId": "githubPullsEnrich",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PullsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GithubPullEnrichmentResult"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "POST /me/github/pulls/enrich — batch-enrich the supplied PR refs.",
        "tags": [
          "github"
        ]
      }
    },
...
    "/api/me/github/repos": {
      "get": {
        "operationId": "githubRepos",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoSelection"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/github/repos — available repos + current selection.",
        "tags": [
          "github"
        ]
      },
      "put": {
        "operationId": "githubSetRepos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReposBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubConnectionSummary"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "PUT /me/github/repos — set the bounded projection scope and bootstrap it.",
        "tags": [
          "github"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Same-origin (the web app proxies /api to this server)",
      "url": "/"
    }
  ],
...
}